+++ tests/test_icepap.py
@@ -1,46 +1,171 @@
 import pytest
+import random
 
-from pyIcePAP.communication import CommType
-from pyIcePAP import EthIcePAPController
+from icepap import IcePAPController
 
 from patch_socket import protect_socket, patch_socket, socket_context
 
 
-def confirm_initial_m1_state(s1):
-    assert s1.status_register == 0x00205013
-    assert s1.is_present()
-    assert s1.is_alive()
-    assert s1.get_mode_str() == 'OPER'
-    assert s1.is_disabled()
-    assert s1.get_disable_str() == 'Axis not active'
+def get_random_pos():
+    return random.randrange(-10000, 10000)
+
+
+def confirm_m1(m1):
+    # Status register 0x00205013
+    assert m1.state_present is True
+    assert m1.state_moving is False
+    assert m1.state_ready is False
+    assert m1.state_settling is False
+    assert m1.state_outofwin is True
+    assert m1.state_warning is False
+    assert m1.state_alive is True
+    assert m1.state_mode_str == 'OPER'
+    assert m1.state_mode_code == 0
+    assert m1.state_disabled is True
+    assert m1.state_disable_code == 1
+    assert m1.state_disable_str == 'Axis not active'
+    assert m1.state_indexer_code == 0
+    assert m1.state_indexer_str == 'Internal indexer'
+    assert m1.state_stop_code == 1
+    assert m1.state_stop_str == 'Stop'
+    assert m1.state_limit_positive is False
+    assert m1.state_limit_negative is False
+    assert m1.state_inhome is False
+    assert m1.state_5vpower is True
+    assert m1.state_verserr is False
+    assert m1.state_poweron is False
+    assert m1.state_info_code == 0
+
+    assert m1.stopcode == 0
+    assert m1.vstopcode == 'No abnormal stop condition'
+    assert m1.active is True
+    assert m1.mode == 'OPER'
+    assert m1.alarm == (False, '')
+    assert m1.config == 'toto@pc1_2019/06/17_12:51:24'
+    assert m1.id == ('0008.028E.EB82', '4960')
+
+    assert m1.meas_vcc == 80.2165
+    assert m1.meas_i == 0.00545881
+    assert m1.meas_ia == -0.00723386
+    assert m1.meas_ib == -0.000653267
+    assert m1.meas_ic == 0
+    assert m1.meas_r == -6894.35
+    assert m1.meas_ra == -3797.74
+    assert m1.meas_rb == -3797.74
+    with pytest.raises(RuntimeError):
+        m1.meas_rc
+
+    pos = get_random_pos()
+    m1.pos = pos
+    assert m1.pos == pos
+    pos = get_random_pos()
+    m1.pos_shftenc = pos
+    assert m1.pos_shftenc == pos
+    pos = get_random_pos()
+    m1.pos_tgtenc = pos
+    assert m1.pos_tgtenc == pos
+    pos = get_random_pos()
+    m1.pos_ctrlenc = pos
+    assert m1.pos_ctrlenc == pos
+    pos = get_random_pos()
+    m1.pos_encin = pos
+    assert m1.pos_encin == pos
+    pos = get_random_pos()
+    m1.pos_inpos = pos
+    assert m1.pos_inpos == pos
+    # TODO implement exception to configuration mode
+    pos = get_random_pos()
+    m1.pos_absenc = pos
+    assert m1.pos_absenc == pos
+    pos = get_random_pos()
+    m1.pos_motor = pos
+    assert m1.pos_motor == pos
+    pos = get_random_pos()
+    m1.pos_sync = pos
+    assert m1.pos_sync == pos
+
+    pos = get_random_pos()
+    m1.enc = pos
+    assert m1.enc == pos
+    pos = get_random_pos()
+    m1.enc_shftenc = pos
+    assert m1.enc_shftenc == pos
+    pos = get_random_pos()
+    m1.enc_tgtenc = pos
+    assert m1.enc_tgtenc == pos
+    pos = get_random_pos()
+    m1.enc_ctrlenc = pos
+    assert m1.enc_ctrlenc == pos
+    pos = get_random_pos()
+    m1.enc_encin = pos
+    assert m1.enc_encin == pos
+    pos = get_random_pos()
+    m1.enc_inpos = pos
+    assert m1.enc_inpos == pos
+    # TODO implement exception to configuration mode
+    pos = get_random_pos()
+    m1.enc_absenc = pos
+    assert m1.enc_absenc == pos
+    pos = get_random_pos()
+    m1.enc_motor = pos
+    assert m1.enc_motor == pos
+    pos = get_random_pos()
+    m1.enc_sync = pos
+    assert m1.enc_sync == pos
+
+    m1.velocity = 200
+    assert m1.velocity == 200
+    assert m1.velocity_current == 200
+    assert m1.velocity_default == 50
+    assert m1.velocity_max == 3000
+    assert m1.velocity_min == 2
+
+    with pytest.raises(AttributeError):
+        m1.velocity_default = 100
+    with pytest.raises(AttributeError):
+        m1.velocity_min = 100
+    with pytest.raises(AttributeError):
+        m1.velocity_max = 100
+
+    m1.acctime = 0.1
+    assert m1.acctime == 0.1
+    assert m1.acctime_default == 0.01
+    assert m1.acctime_steps == 30
+
+    assert m1.pcloop is True
+    m1.pcloop = False
+    assert m1.pcloop is False
+
+    assert m1.indexer == 'INTERNAL'
 
 
 def ice_auto_axes(f):
     """A helper which provides parametrized auto_axes version of icepap"""
-    @pytest.mark.parametrize('auto_axes', [True, False], ids=['smart', 'expert'])
+    @pytest.mark.parametrize('auto_axes', [True, False],
+                             ids=['smart', 'expert'])
     def wrapper(auto_axes):
         with socket_context() as mock_sock:
             patch_socket(mock_sock)
-            pap = EthIcePAPController('icepap1', auto_axes=auto_axes)
+            pap = IcePAPController('icepap1', auto_axes=auto_axes)
             return f(pap)
     return wrapper
 
 
-@pytest.mark.parametrize('auto_axes', [True, False], ids=['smart', 'expert'])
+@pytest.mark.parametrize('auto_axes', [True, False],
+                         ids=['smart', 'expert'])
 def test_create_eth_icepap(auto_axes):
     with socket_context() as mock_sock:
         patch_socket(mock_sock)
         with pytest.raises(RuntimeError):
-            ice = EthIcePAPController('weirdhost')
+            IcePAPController('weirdhost')
         with pytest.raises(RuntimeError):
-            ice = EthIcePAPController('icepap1', 5001)
-        ice = EthIcePAPController('icepap1', 5000)
+            IcePAPController('icepap1', 5001)
+        ice = IcePAPController('icepap1', 5000)
         assert ice is not None
 
 
 @ice_auto_axes
 def test_connection(pap):
-    assert pap.comm_type == CommType.Socket
     assert pap.connected
 
     pap.disconnect()
@@ -50,7 +175,9 @@ def test_connection(pap):
 @ice_auto_axes
 def test_system(pap):
     assert pap.mode == 'OPER'
-
+    m1 = pap[1]
+    confirm_m1(m1)
+    m1.pos = 55
     assert pap.get_pos(1) == [55]
     assert pap.get_pos([1]) == [55]
     assert pap.get_pos([1, 5]) == [55, -3]
@@ -69,11 +196,9 @@ def test_system(pap):
     assert pap.get_fstatus([1]) == [0x00205013]
     assert pap.get_fstatus([1, 5]) == [0x00205013, 0x00205013]
 
-    s1 = pap.get_states(1)[0]
-    confirm_initial_m1_state(s1)
     assert pap.get_states([1])[0].status_register == 0x00205013
-    assert [s.status_register for s in pap.get_states([1, 5])] == [
-                                                            0x00205013, 0x00205013]
+    assert [s.status_register for s in pap.get_states([1, 5])] == \
+           [0x00205013, 0x00205013]
 
     assert pap.get_power(1) == [True]
     assert pap.get_power([1]) == [True]
@@ -141,12 +266,10 @@ def test_smart_axis(smart_pap):
 
     assert m1.addr == 1
     assert m1.name == 'th'
-    assert m1.pos == 55
     assert m1.status == 0x00205013
-    assert m1.power == True
-    s1 = m1.state
-    confirm_initial_m1_state(s1)
-
+    assert m1.power is True
+    confirm_m1(m1)
+    m1.pos = 55
     assert smart_pap.get_pos(1) == [55]
     assert smart_pap.get_pos('toto') == [55]
 
@@ -194,11 +317,10 @@ def test_expert_axis(expert_pap):
 
     assert m1.addr == 1
     assert m1.name == 'th'
-    assert m1.pos == 55
     assert m1.status == 0x00205013
-    assert m1.power == True
-    s1 = m1.state
-    confirm_initial_m1_state(s1)
+    assert m1.power is True
+    confirm_m1(m1)
+    m1.pos = 55
 
     assert expert_pap.get_pos(1) == [55]
     assert expert_pap.get_pos('toto') == [55]